Fix: 모달창 오류 수정 (작업중)
